backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15
drf-spectacular>=0.27
django-cors-headers>=4.3
django-filter>=24.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
celery>=5.3
Pillow>=10.0
python-decouple>=3.8
gunicorn>=21.0
whitenoise>=6.6
requests>=2.31
reportlab>=4.1
supabase>=2.3
